can define and power pin delete log

+++ USER/can_speed.h
@@ -3,11 +3,11 @@
 
 #include "stm32f10x.h"
 
-extern volatile u8 g_speed;
+extern volatile u16 g_speed;
 
-void uart_send_speed(u8 speed);
+void uart_send_speed(u16 speed);
 
-void can_send_speed(u8 speed);
+void can_send_speed(u16 speed);
 
 void can_resolve_speed(u8 *buf,u8 idex);
 
@@ -15,6 +15,6 @@ s8 can_obdii_resolve_speed(u8 *buf);
 
 s8 obdii_requeset(int pid);
 
-u8 get_can_speed(void);
+u16 get_can_speed(void);
 
 #endif
