+++ USER/speed_send_task.c
@@ -6,8 +6,6 @@
 
 void speed_send_task(OS_TMR *ptmr, void *pdata)
 {
-		 u8 speed =get_can_speed();	 
-		 //can_send_speed(speed);
+		 u16 speed =get_can_speed();
 		 uart_send_speed(speed);
 }
-
